requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.5.1
certifi==2022.5.18.1
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.7.2
python-telegram-bot==13.3
pytz==2021.1
requests==2.26
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==1.0.1
wrapt==1.12.1
gTTS==2.2.4
yfinance==0.1.70
mplfinance==0.12.8b9
python-decouple==3.6
path-16.4.0